backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
PyMuPDF==1.27.2
pytesseract==0.3.13
Pillow==12.1.1
python-docx==1.2.0
django-ratelimit==4.1.0
django-cors-headers==4.9.0
openai==2.26.0
djangorestframework-simplejwt==5.5.0
google-generativeai==0.8.2
